adds BinaryQuestion component
